## PR Summary

As pointed out by @anntzer constrained layout gets slow for many subplots.  Some slowdown is to be expected, but there were numerous redundant constraints being added, so this now has watchdogs to make sure only one constraint is added.  Also some (probably insignificant) optimizations to reduce the number of calculations.  

"closes" #10655 

- subplots(4, 4) took    4s    now   1.8s  
- subplots(5, 5) took  12s,   now   3.3s
- subplots(6, 6) took  30s,  now   6.3s

Still not quite linear in N, but a lot better than N*2 which it was before.  

```python
import matplotlib.pyplot as plt
import timeit

start = timeit.default_timer()

fig, ax = plt.subplots(6, 6, constrained_layout=False)
fig.canvas.draw()
stop = timeit.default_timer()
print(stop-start)
```

## PR Checklist

- [ ] Has Pytest style unit tests
- [ ] Code is PEP 8 compliant
- [ ] New features are documented, with examples if plot related
- [ ] Documentation is sphinx and numpydoc compliant
- [ ] Added an entry to doc/users/next_whats_new/ if major new feature (follow instructions in README.rst there)
- [ ] Documented in doc/api/api_changes.rst if API changed in a backward-incompatible way

<!--
Thank you so much for your PR!  To help us review your contribution, please
consider the following points:

- A development guide is available at https://matplotlib.org/devdocs/devel/index.html.

- Help with git and github is available at
  https://matplotlib.org/devel/gitwash/development_workflow.html.

- Do not create the PR out of master, but out of a separate branch.

- The PR title should summarize the changes, for example "Raise ValueError on
  non-numeric input to set_xlim".  Avoid non-descriptive titles such as
  "Addresses issue #8576".

- The summary should provide at least 1-2 sentences describing the pull request
  in detail (Why is this change required?  What problem does it solve?) and
  link to any relevant issues.

- If you are contributing fixes to docstrings, please pay attention to
  http://matplotlib.org/devel/documenting_mpl.html#formatting.  In particular,
  note the difference between using single backquotes, double backquotes, and
  asterisks in the markup.

We understand that PRs can sometimes be overwhelming, especially as the
reviews start coming in.  Please let us know if the reviews are unclear or
the recommended next step seems overly demanding, if you would like help in
addressing a reviewer's comments, or if you have been waiting too long to hear
back on your PR.
-->