@jklymak 2018-03-03
- subplots(4, 4) took    4s    now   1.8s  
- subplots(5, 5) took  12s,   now   3.3s
- subplots(6, 6) took  30s,  now   6.3s

Still not quite linear in N, but a lot better than worse than N*2 which it was before.