@jklymak 2018-03-03
This need not go out tomorrow (or whenever 2.2 goes out), but it would be nice to speed this up sooner rather than later.